email sending fixed.
Not related to the email stuff: SmoothScroll.js had an error in a
verification routine of some sort that I don't understand. I
commented the whole routine so that the user doesn't see any errors
on the browser console.

I still want to check the email parts because the address bar becomes
cluttered with junk from the request.